Updated post to include data from document and updated tests for that data. Added web config to remove cors

+++ rev-connect-api/src/main/java/com/rev_connect_api/repositories/PostRepository.java
@@ -4,12 +4,13 @@ import com.rev_connect_api.models.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post, Integer> {
 
-    Optional<Post> getPostByPostId(int id);
+    Optional<Post> getPostByPostId(BigInteger id);
 
-    void deletePostByPostId(int id);
+    void deletePostByPostId(BigInteger id);
 }
